Create soul_soil.json (#2168)
* Create soul_soil.json

Copy of the centrifuge soul_sand recipe, but for soul_soil
I'm sorry, I don't know if I need to add a reference to this recipe anywhere else.

* Update soul_soil.json

Adjusted the recipe to return 8 minecraft:dirt instead of 10 minecraft:sand.  Lowered the amount of dirt, because it's more dynamic than sand for processing it further.  According to REI, this would be the first TR recipe to provide dirt. (besides the scrapbox)