get drainFluid via ItemFluidInfo
Previously FluidDrainable.tryDrainFluid would return a Fluid, but now it
returns an ItemStack (in vanilla it's always a BucketItem). Vanilla has
no proper way of retrieving the Fluid contained in the ItemStack, but we
can use RebornCore's ItemFluidInfo, which should also work for other
Fluid-containing Items like Cells, if they are returned by tryDrainFluid.