Final commit to resolve #2675 I hope. Thanks to Ayutac
Bug history: Cells were able to put out fluids which are not FlowableFluids, causing a textureless block to appear. First commit checked for FlowableFluid necessary in cell item to do anything with it. However since EMPTY is not a FlowableFluid that led to Cells being unable to pick up fluids any longer. Second commit attempted to fix but didn't because I checked it wrongly (tested with buckets instead of cells). In fact it doesn't do anything (because the same check was performed earlier), so I reverted it with this commit. And finally this commit adds EMPTY as an acceptable fluid inside a cell to do something with it.