Fixed AutoCrafting Table bug
If the table was locked, canMake only checked for each ingredient to be there more than once, but make greedily took all items from the first encountered slot of the needed ingredient if the recipe didn't use the full crafting table and the ingredients weren't aligned to the upper left corner of the crafting table (discovered and reproducible with three wheat in the first column second row and single wheats in the other two columns of the second row).

This commit simply disallows greedy behaviour, each crafting grid slot can only be used once.

I also made one or two changes to the rest of the code (formatting,use of constants etc) while I was at it, but these didn't change functionality.