Small Piglin tag update (#2770)
* Small Piglin tag update

added gold plate and electrum to piglin stuff. Piglins should be able to smell the gold out of an electrum ingot and a block, but not a nugget. Baby piglins are worse at this.

Explicitly used TR IDs instead of c tags because this is part of TR flavor imo

* forgot electrum plate

* and gold cables