Change Primary output to Raw Iridium
Changes primary output to 2x raw iridium as other recipes that use mercury with this machine generally have a greater output than just with water.
Makes more sense to have the raw ore as the output from this machine instead of an ingot, which you would usually get from a furnace of some type.