In-depth update of chrome->chromium, thanks to Ayutac, Xanthian
* In-depth update of chrome->chromium, thanks to Ayutac, Xanthian

This fixes Bug #2621 without accidentally destroying items (untested, but the chrome item registration isn't changed) and supersedes PR #2780 which only made language changes.

I just plainly copied the lang changes from Xanthian. As for the chrome stuff, I thought long and hard how to do it properly. Possible variants:
1. Change tag names in constructor for chrome only.
2. Change tag names in enum for chrome only.
3. Don't change the internal tag names of the enums, but change the tag registration process in datagen for chrome only.
4. Introduce CHROMIUM variant, add conversion recipes etc. and hope over time CHROME can be safely removed.

(4) always had the potential of item loss, so I discarded it first. Deciding between (1), (2) and (3) was harder. The drawback of (2) is code bloating by adding a lot of extra enum constructors, especially for cases which already had enum constructor options. But in the end I decided to go with it because it was the cleanest solution to a dirty problem. This is the first time I encountered a datagen drawback, because if we didn't register by running through the enum array, we could have simply changed it there; a possible (5) but still worse than (2).

* Whoops, forgot recipe changes with the tags

* Merged chromium with hardness/resistance

Replaced the isHot argument from constructors with default false while I was at it.